+++ js/create-advert.js
@@ -1,0 +1,9 @@
+const getAdvert = () => {
+  const locationPoint = getLocationPoint();
+  return {
+    author: getAuthor(),
+    offer: getOffer(locationPoint),
+    location: locationPoint
+  };
+};
+export {getAdvert};
