+++ js/create-adverts-list.js
@@ -1,0 +1,5 @@
+const getAdvertsList = () => Array.from({ length: AMOUNT_USERS }, getAdvert);
+
+getAdvertsList();
+
+export {getAdvertsList};
