+++ js/create-links.js
@@ -1,0 +1,8 @@
+
+const createRandomLinksId = createRandomIdFromRangeGenerator(1, AMOUNT_USERS);
+
+const getRandomLinksId = () => {
+  const currentLinksId = createRandomLinksId();
+  return (currentLinksId < AMOUNT_USERS) ? `${0}${currentLinksId}` : currentLinksId;
+};
+export {getRandomLinksId};
