+++ js/create-location.js
@@ -1,0 +1,5 @@
+const getLocationPoint = () => ({
+  lat: getRandomPoint(35.65000, 35.70000, 5),
+  lng: getRandomPoint(139.70000, 139.80000, 5),
+});
+export {getLocationPoint};
