+++ js/create-user.js
@@ -1,0 +1,4 @@
+const getAuthor = () => ({
+  avatar: `img/avatars/user${getRandomLinksId()}.png`
+});
+export {getAuthor};
