+++ js/data.js
@@ -1,0 +1,10 @@
+const AMOUNT_USERS = 10;
+
+const typesAccommodation = ['palace', 'flat', 'house', 'bungalow', 'hotel'];
+const checkTimes = ['12:00', '13:00', '14:00'];
+const facilities = ['wifi', 'dishwasher', 'parking', 'washer', 'elevator', 'conditioner'];
+const photoLinks = ['https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg'];
+
+export {AMOUNT_USERS, typesAccommodation, checkTimes, facilities, photoLinks};
